moved tool code into toolcmds
